# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Online Examination Platform
  version: 1.1.0
  description: >-
    Pre-release public contract amended by ADR-005 for verified email and asymmetric session JWTs.
    No runtime deployment. Product/security/SLO requirements in docs; all
    response examples are fixtures.
...
servers:
  - url: /
    description: Same HTTPS viewer origin; account/region/domain/TLS deployment pending
...
paths:
...
  /v1/exams/{examId}/attempts:
    post:
      operationId: startAttempt
      summary: Start current version or return existing active attempt without resetting time
      tags:
        - Assessment
      security:
        - AccessCookie: []
          CsrfHeader: []
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
      responses:
        "201":
          description: Durable accepted outcome
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
            Idempotency-Replayed:
              description: true for replay of durable original receipt.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptEnvelope"
              example:
                data:
                  id: 00000000-0000-4000-8000-000000000004
                  examId: 00000000-0000-4000-8000-000000000001
                  publishedVersionId: 00000000-0000-4000-8000-000000000002
                  revision: 1
                  status: IN_PROGRESS
                  startedAt: "2026-10-06T13:00:00.000Z"
                  deadline: "2026-10-06T13:45:00.000Z"
                  submittedAt: null
                  expired: false
                  serverNow: "2026-10-06T13:10:00.000Z"
                  canSave: true
                  resultAvailable: false
                  pollAfterSeconds: 0
                  replayPending: false
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: write
      x-permissions:
        - assessment.take
      x-transaction: start+receipt
      x-idempotency: true
      x-cases:
        - AC-03
        - AC-04
        - AC-05
        - AC-31
      x-query-budget: 12
  /v1/attempts/{attemptId}:
    get:
      operationId: resumeAttempt
      summary: resumeAttempt
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptEnvelope"
              example:
                data:
                  id: 00000000-0000-4000-8000-000000000004
                  examId: 00000000-0000-4000-8000-000000000001
                  publishedVersionId: 00000000-0000-4000-8000-000000000002
                  revision: 1
                  status: IN_PROGRESS
                  startedAt: "2026-10-06T13:00:00.000Z"
                  deadline: "2026-10-06T13:45:00.000Z"
                  submittedAt: null
                  expired: false
                  serverNow: "2026-10-06T13:10:00.000Z"
                  canSave: true
                  resultAvailable: false
                  pollAfterSeconds: 0
                  replayPending: false
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.take
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-04
        - AC-05
        - AC-18
      x-query-budget: 3
  /v1/attempts/{attemptId}/questions:
    get:
      operationId: getQuestions
      summary: getQuestions
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateQuestionListEnvelope"
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.take
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-02
        - AC-18
        - AC-19
        - AC-27
      x-paginated: true
      x-query-budget: 4
  /v1/attempts/{attemptId}/answers:
    get:
      operationId: getAnswers
      summary: getAnswers
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerListEnvelope"
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.take
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-06
        - AC-18
        - AC-27
      x-paginated: true
      x-query-budget: 4
    put:
      operationId: saveAnswers
      summary: saveAnswers
      tags:
        - Assessment
      security:
        - AccessCookie: []
          CsrfHeader: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: Durable accepted outcome
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
            Idempotency-Replayed:
              description: true for replay of durable original receipt.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveReceiptEnvelope"
              example:
                data:
                  attemptId: 00000000-0000-4000-8000-000000000004
                  acceptedAt: "2026-10-06T13:10:00.000Z"
                  answers:
                    - questionId: 00000000-0000-4000-8000-000000000010
                      version: 1
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: write
      x-permissions:
        - assessment.take
      x-transaction: attempt-lock+answers+receipt
      x-idempotency: true
      x-cases:
        - AC-06
        - AC-07
        - AC-08
        - AC-09
        - AC-10
        - AC-31
      x-query-budget: 11
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnswersRequest"
            example:
              answers:
                - questionId: 00000000-0000-4000-8000-000000000010
                  selectedOptionIds:
                    - 00000000-0000-4000-8000-000000000003
                  marked: false
                  expectedVersion: 0
  /v1/attempts/{attemptId}/submit:
    post:
      operationId: submitAttempt
      summary: submitAttempt
      tags:
        - Assessment
      security:
        - AccessCookie: []
          CsrfHeader: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Origin"
      responses:
        "202":
          description: Durable accepted outcome
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
            Idempotency-Replayed:
              description: true for replay of durable original receipt.
              schema:
                type: boolean
            Retry-After:
              description: Seconds before retry/poll.
              schema:
                type: integer
                minimum: 1
                maximum: 60
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitReceiptEnvelope"
              example:
                data:
                  attemptId: 00000000-0000-4000-8000-000000000004
                  submissionId: 00000000-0000-4000-8000-000000000005
                  acceptedAt: "2026-10-06T13:44:50.000Z"
                  acceptanceState: SUBMITTED
                  expired: false
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: write
      x-permissions:
        - assessment.take
      x-transaction: attempt-lock+submission+outbox+receipt
      x-idempotency: true
      x-cases:
        - AC-09
        - AC-10
        - AC-11
        - AC-31
      x-query-budget: 10
  /v1/attempts/{attemptId}/status:
    get:
      operationId: getAttemptStatus
      summary: getAttemptStatus
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptEnvelope"
              example:
                data:
                  id: 00000000-0000-4000-8000-000000000004
                  examId: 00000000-0000-4000-8000-000000000001
                  publishedVersionId: 00000000-0000-4000-8000-000000000002
                  revision: 1
                  status: IN_PROGRESS
                  startedAt: "2026-10-06T13:00:00.000Z"
                  deadline: "2026-10-06T13:45:00.000Z"
                  submittedAt: null
                  expired: false
                  serverNow: "2026-10-06T13:10:00.000Z"
                  canSave: true
                  resultAvailable: false
                  pollAfterSeconds: 0
                  replayPending: false
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.take
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-11
        - AC-17
        - AC-18
      x-query-budget: 3
  /v1/attempts/{attemptId}/result:
    get:
      operationId: getResult
      summary: getResult
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultEnvelope"
              example:
                data:
                  attemptId: 00000000-0000-4000-8000-000000000004
                  publishedVersionId: 00000000-0000-4000-8000-000000000002
                  submissionId: 00000000-0000-4000-8000-000000000005
                  completedAt: "2026-10-06T13:45:10.000Z"
                  earned: 5
                  possible: 6
                  correct: 2
                  total: 3
                  percentageBasisPoints: 8333
                  expired: false
                  scoringPolicy: EXACT_MATCH_V1
                  sections:
                    - sectionId: 00000000-0000-4000-8000-000000000011
                      earned: 5
                      possible: 5
                      correct: 2
                      total: 2
                    - sectionId: 00000000-0000-4000-8000-000000000012
                      earned: 0
                      possible: 1
                      correct: 0
                      total: 1
                  review: null
                errorCode: null
                message: null
                status: true
        "202":
          description: Not completed; durable status, no fabricated score
          headers:
            Retry-After:
              description: Seconds before retry/poll.
              schema:
                type: integer
                minimum: 1
                maximum: 60
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptEnvelope"
              example:
                data:
                  id: 00000000-0000-4000-8000-000000000004
                  examId: 00000000-0000-4000-8000-000000000001
                  publishedVersionId: 00000000-0000-4000-8000-000000000002
                  revision: 1
                  status: IN_PROGRESS
                  startedAt: "2026-10-06T13:00:00.000Z"
                  deadline: "2026-10-06T13:45:00.000Z"
                  submittedAt: null
                  expired: false
                  serverNow: "2026-10-06T13:10:00.000Z"
                  canSave: true
                  resultAvailable: false
                  pollAfterSeconds: 0
                  replayPending: false
                errorCode: null
                message: null
                status: true
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.result.read
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-17
        - AC-18
        - AC-19
      x-query-budget: 3
  /v1/attempts/{attemptId}/review:
    get:
      operationId: getReleasedReview
      summary: Completed owned attempt and frozen explanation release gate
      description: Cursor page bounded to 100 items and 256KiB complete decoded JSON; a payload-limited page may contain fewer than pageSize items. Follow metadata.next.
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQuestionListEnvelope"
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.result.read
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-18
        - AC-19
        - AC-27
      x-paginated: true
      x-query-budget: 3
  /v1/me/attempts:
    get:
      operationId: getHistory
      summary: getHistory
      tags:
        - Assessment
      security:
        - AccessCookie: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Authorized projection
          headers:
            X-Correlation-Id:
              description: Server-generated/validated UUID correlation ID.
              schema:
                type: string
                format: uuid
            Cache-Control:
              description: no-store for authenticated/sensitive payloads.
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryItemListEnvelope"
        "400":
          $ref: "#/components/responses/Http400"
        "401":
          $ref: "#/components/responses/Http401"
        "403":
          $ref: "#/components/responses/Http403"
        "404":
          $ref: "#/components/responses/Http404"
        "409":
          $ref: "#/components/responses/Http409"
        "422":
          $ref: "#/components/responses/Http422"
        "429":
          $ref: "#/components/responses/Http429"
        "500":
          $ref: "#/components/responses/Http500"
        "503":
          $ref: "#/components/responses/Http503"
      x-owner: Assessment
      x-flow: read
      x-permissions:
        - assessment.take
      x-transaction: none
      x-idempotency: false
      x-cases:
        - AC-18
        - AC-27
      x-paginated: true
      x-query-budget: 3
...
components:
  securitySchemes:
    AccessCookie:
      type: apiKey
      in: cookie
      name: __Host-access
      description: Asymmetric signed access JWT; strict purpose validation plus authoritative DB session/permissions check.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-refresh
      description: Asymmetric signed refresh JWT; single-use rotation and durable family revocation in PostgreSQL.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Signed cookie echo plus exact Origin validation; not authentication.
  responses:
    Http400:
      description: Invalid request
      headers:
        X-Correlation-Id:
          description: Server-generated/validated UUID correlation ID.
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            data: null
            errorCode: Invalid request
            message: Invalid request
            status: false
    Http401:
      description: Unauthenticated
      headers:
        X-Correlation-Id:
          description: Server-generated/validated UUID correlation ID.
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            data: null
            errorCode: Unauthenticated
            message: Unauthenticated
            status: false
    Http403:
      description: Permission denied
      headers:
        X-Correlation-Id:
          description: Server-generated/validated UUID correlation ID.
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            data: null
            errorCode: Permission denied
            message: Permission denied
            status: false
...
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      description: >-
        Exact allowlisted HTTPS viewer origin; browser supplies it. Local development uses explicitly
        configured localhost origin.
      schema:
        type: string
        format: uri
        maxLength: 512
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Canonical UUIDv7, first use within 24h and <=5min future skew; compact receipt retained >=7days.
        ADR-003.
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 2048
      description: Signed keyset cursor bound to filter/version/order/actor/page size; expires 15min.
...
  schemas:
...
    AttemptEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - errorCode
        - message
        - status
      properties:
        data:
          $ref: "#/components/schemas/Attempt"
        errorCode:
          type: string
          nullable: true
          enum: [null]
        message:
          type: string
          nullable: true
          enum: [null]
        status:
          type: boolean
          enum:
            - true
...